Binary search code
searching is used to find the location where an element is available . There are two types of search techniques 
1) Linear search 
2) Binary search

Linear search:   This is simplest of all searching techniques. In this technique, an ordered or unordered list will be searched one by one from the beginning until the desire element is found. If the desired element is found in the list, then the search is successful otherwise unsuccessful.

How Does Linear Search Algorithm Work?
In Linear Search Algorithm, 

Every element is considered as a potential match for the key and checked for the same.
If any element is found equal to the key, the search is successful and the index of that element is returned.
If no element is found equal to the key, the search yields “No match found”.


Algorithm
Linear_Search(a, n, val) // 'a' is the given array, 'n' is the size of given array, 'val' is the value to search.  
Step 1: set pos = -1  
Step 2: set i = 1  
Step 3: repeat step 4 while i <= n  
Step 4: if a[i] == val  
set pos = i  
print pos  
go to step 6  
[end of if]  
set ii = i + 1  
[end of loop]  
Step 5: if pos = -1  
print "value is not present in the array "  
[end of if]  
Step 6: exit