Linear Search Code
searching is used to find the location where an element is available . There are two types of search techniques 
1) Linear search 
2) Binary search

Linear search:   This is simplest of all searching techniques. In this technique, an ordered or unordered list will be searched one by one from the beginning until the desire element is found. If the desired element is found in the list, then the search is successful otherwise unsuccessful.

How Does Linear Search Algorithm Work?
In Linear Search Algorithm, 

Every element is considered as a potential match for the key and checked for the same.
If any element is found equal to the key, the search is successful and the index of that element is returned.
If no element is found equal to the key, the search yields “No match found”.


Algorithm
Linear_Search(a, n, val) // 'a' is the given array, 'n' is the size of given array, 'val' is the value to search.  
Step 1: set pos = -1  
Step 2: set i = 1  
Step 3: repeat step 4 while i <= n  
Step 4: if a[i] == val  
set pos = i  
print pos  
go to step 6  
[end of if]  
set ii = i + 1  
[end of loop]  
Step 5: if pos = -1  
print "value is not present in the array "  
[end of if]  
Step 6: exit  


Binary Search
Binary Search is a searching algorithm for finding an element's position in a sorted array.
In this approach, the element is always searched in the middle of a portion of an array.

The general steps for both methods are discussed below.

The array in which searching is to be performed is:
initial array Binary Search
Initial array

Let x = 4 be the element to be searched.
Set two pointers low and high at the lowest and the highest positions respectively.
setting pointers Binary Search
Setting pointers
Find the middle element mid of the array ie. arr[(low + high)/2] = 6.
mid element Binary Search
Mid element
If x == mid, then return mid.Else, compare the element to be searched with m.
If x > mid, compare x with the middle element of the elements on the right side of mid. This is done by setting low to low = mid + 1.
Else, compare x with the middle element of the elements on the left side of mid. This is done by setting high to high = mid - 1.
finding mid element Binary Search
Finding mid element
Repeat steps 3 to 6 until low meets high.
mid element Binary Search
Mid element
x = 4 is found.
found Binary Search
Found
Binary Search Algorithm
Iteration Method
do until the pointers low and high meet each other.
    mid = (low + high)/2
    if (x == arr[mid])
        return mid
    else if (x > arr[mid]) // x is on the right side
        low = mid + 1
    else                       // x is on the left side
        high = mid - 1